chore: secure org and tx api routes

+++ pages/api/organizations/[id].ts
@@ -1,16 +1,24 @@
 import type { NextApiRequest, NextApiResponse } from "next";
+import { unstable_getServerSession } from "next-auth";
 import { prisma } from "../../../src/lib/db";
+import { authOptions } from "../auth/[...nextauth]";
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
+  const session = await unstable_getServerSession(req, res, authOptions(req));
+
+  if (!session?.user?.name) {
+    return res.status(401).json({ message: "Unauthorized" });
+  }
+
   switch (req.method) {
     case "GET":
-      await handleGetOrganization(req, res);
+      await handleGetOrganization(req, res, session.user.name);
       break;
     case "PATCH":
-      await handleEditOrganization(req, res);
+      await handleEditOrganization(req, res, session.user.name);
       break;
     case "DELETE":
-      await handleDeleteOrganization(req, res);
+      await handleDeleteOrganization(req, res, session.user.name);
       break;
     default:
       res.status(405).json({ message: "Method not allowed" });
@@ -19,7 +27,8 @@ const handler = async (req: NextApiRequest, res: NextApiResponse) => {
 
 const handleGetOrganization = async (
   req: NextApiRequest,
-  res: NextApiResponse
+  res: NextApiResponse,
+  pubkey: string
 ) => {
   if (!req.query.id) {
     return res.status(400).json({ message: "Bad request" });
@@ -30,7 +39,20 @@ const handleGetOrganization = async (
       where: {
         id: req.query.id as string,
       },
+      include: {
+        members: {
+          where: {
+            profile: {
+              pubkey,
+            },
+          },
+        },
+      },
     });
+
+    if (!organization) {
+      return res.status(404).json({ message: "Not found" });
+    }
 
     return res.status(200).json({ organization });
   } catch (error) {
@@ -41,14 +63,34 @@ const handleGetOrganization = async (
 
 const handleEditOrganization = async (
   req: NextApiRequest,
-  res: NextApiResponse
+  res: NextApiResponse,
+  pubkey: string
 ) => {
   if (!req.query.id) {
     res.status(400).json({ message: "Bad request" });
   }
 
   try {
-    const organization = await prisma.organization.update({
+    const organization = await prisma.organization.findUnique({
+      where: {
+        id: req.query.id as string,
+      },
+      include: {
+        members: {
+          where: {
+            profile: {
+              pubkey,
+            },
+          },
+        },
+      },
+    });
+
+    if (!organization) {
+      return res.status(404).json({ message: "Not found" });
+    }
+
+    const updatedOrganization = await prisma.organization.update({
       where: {
         id: req.query.id as string,
       },
@@ -61,7 +103,7 @@ const handleEditOrganization = async (
       },
     });
 
-    return res.status(200).json({ organization });
+    return res.status(200).json({ organization: updatedOrganization });
   } catch (error) {
     return res.status(500).json({ message: "Internal server error", error });
   }
@@ -69,20 +111,40 @@ const handleEditOrganization = async (
 
 const handleDeleteOrganization = async (
   req: NextApiRequest,
-  res: NextApiResponse
+  res: NextApiResponse,
+  pubkey: string
 ) => {
   if (!req.query.id) {
     res.status(400).json({ message: "Bad request" });
   }
 
   try {
-    const organization = await prisma.organization.delete({
+    const organization = await prisma.organization.findUnique({
+      where: {
+        id: req.query.id as string,
+      },
+      include: {
+        members: {
+          where: {
+            profile: {
+              pubkey,
+            },
+          },
+        },
+      },
+    });
+
+    if (!organization) {
+      return res.status(404).json({ message: "Not found" });
+    }
+
+    const deletedOrganization = await prisma.organization.delete({
       where: {
         id: req.query.id as string,
       },
     });
 
-    return res.status(200).json({ organization });
+    return res.status(200).json({ organization: deletedOrganization });
   } catch (error) {
     return res.status(500).json({ message: "Internal server error", error });
   }
